Sistema_verifica_documentale_appalti/requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
Pillow
python-multipart
